Upgrade to Angular 16
This commit includes the package.json and lockfile updates made to
upgrade Angular to version 16.

These changes were made from the following command:
```
npx ng update @angular-eslint/builder@16 \
    @angular-eslint/eslint-plugin@16 \
    @angular-eslint/eslint-plugin-template@16 \
    @angular-eslint/schematics@16 \
    @angular-eslint/template-parser@16 \
    @angular/core@16 @angular/cli@16 @angular/cdk@16 \
    @angular/google-maps@16 @angular/compiler-cli@16 \
    @angular/language-service@16 @angular/localize@16 \
    @angular/animations@16 @angular/common@16 \
    @angular/forms@16 @angular/platform-browser@16 \
    @angular/router@16 @ng-bootstrap/ng-bootstrap@15 \
    shallow-render@16 @angular-devkit/build-angular@16 \
    @fortawesome/angular-fontawesome@^0.13 angular-shepherd@16 \
    ngx-cookie-service@16 ng-recaptcha@^12 ts-jest@29
```

Missing is an upgrade of the dependency @angular/fire to version 16.
This is because this package changed how versioning of their releases is
done to match the major version of the Angular version they are
compatible with. As a result, @angular/fire went from version 7.x to
version 16.0.0 in one release. The Angular CLI refuses to allow upgrades
of more than one major version at a time, even if @angular/fire versions
8 through 15 do not exist. This will be done manually in an upcoming
commit.

PER-9399: Upgrade to Angular 16