Create RouteHistoryModule
The web-app currently relies on a package called ngx-route-history for
some functionality. Unfortunately, this package does not seem to be
maintained anymore and it is incompatible with Angular 16. Since this
was a custom package made by a previous developer, it makes more sense
for the web-app to fully own this code so it's always compatible.

Create the RouteHistoryModule with the RouteHistory service, and import
the tests from the original package. The tests were modified to use real
async instead of the `fakeAsync` that was used in older Angular tests,
as well and the `ngOnDestroy` test was rewritten as well to be simpler.

The service itself was not really adapted from the original, instead
developed with Test Driven Development using the existing tests.
Although the service is so simple that the implementations end up being
very similar.

PER-9399: Upgrade to Angular 16