Use the new RouteHistoryModule
Use the new RouteHistoryModule in the application and point to the
locally owned RouteHistoryService instead of the one that's provided by
the ngx-route-history package.

There is a slight architectural change in this new implementation: the
RouteHistoryService is now contained in its own module that provides it
and this module is imported by the AppModule. This is in contrast to the
original implementation, in which the RouteHistoryService was
essentially a standalone service that manually had to be provided by the
AppModule.

This does not affect any functionality and probably sets up better
boundaries in our code now that the RouteHistoryService is locally
owned.

PER-9399: Upgrade to Angular 16