Don't redirect invalid session on Public Gallery
Fix a bug where an invalid session on the public gallery could redirect
the user to the login screen. Since authentication is not needed in the
public gallery, they do not need to be redirected.

While this fixes some bug that deals with redirections to login, it's
unclear if this is actually fixing the bug referred to in PER-9409. The
invalid session redirect is one of the few pathways available for
redirection to the login screen on the public gallery, and it seems to
be the only one that does not provide an error message as seen in the
one video recording we have of the bug in action. However, it's unclear
how this specific code would be triggered in a brand new browser that
doesn't have any session data in it to begin with, which is what was
observed in the video.

To keep the AccountService tests more organized, the tests for this
change were put in a separate file since the main AccountService spec
file was already getting very long. Because of the way Angular unit
testing works, there's a limit to how much code duplication we can
eliminate, so we may have to split additional tests into their own files
to keep things more organized.

PER-9409: Public archive sometimes redirects you to the login