Fetch Featured Archives from Stela
Continuing from previous work, create a service that implements the
`FeaturedArchiveApi` interface. Ultimately I decided to let this code
live fully within the GalleryModule instead of adding it to the
ApiService. The main reason for doing so is that this code will only
ever be called within the GalleryModule; no other parts of the codebase
currently need to know what a `FeaturedArchive` is and no other parts of
the application will ever need to hit the `archive/featured` endpoint,
so there's no reason to add these types and methods to the shared
ApiService.