Add basic SearchBoxComponent test
The SearchBoxComponent needed one of its properties to be renamed as
part of the work on Featured Archives. However, since SearchBoxComponent
had no tests, this change was not considered covered by tests. Fix this
by creating a basic "it exists" test for the SearchBoxComponent. Ideally
we'd flesh this out to increase coverage in this component as well, but
that work can be done when we have actual functional adjustments that
need to be made to the component. Until then, having this basic test is
much better than having nothing at all.