Make `new-line-before-expect` rule throw error
Angular doesn't seem to have a way for eslint to be more strict when
running lint checks. The `new-line-before-expect` rule recently added
defaults to throwing warnings in eslint, which does not register as a
lint failure in Github Actions. Change it to throw errors instead so
that we start using this rule consistently.