Refactor timeline-view type fixes
Use a helper method to convert the DateType objects in a more type safe
way. While we are getting them as numbers in the actual application,
it's better to prepare for them being Date or String objects as well.
Write a helper method to deal with all 3 types that DateType can be.