Create PermanentFilesystem class
This class is intended to be a new central location for fetching parts
of the Permanent filesystem from the API. Initially the class is set up
just to wrap around an interface, with the idea that caching logic will
be built around this code. Eventually when connected to the rest of the
application, any sites that call the API methods will be replaced with
calls to these methods instead.

PER-9453: "Scaffolding for caching system"