Update Profile/FolderPicker service injection type
This configuration was made to get tests to pass but created issues
where there were suddenly *two* instances of FolderPickerService for
both the CoreModule and the SharedModule. Fix this by reverting the
Injectable decorator in both the ProfileService and FolderPickerService
and also put the ProfileService in its own module for testing purposes.

PER-9471: Buttons to change profile and banner photo not working