Implement folder caching in PermanentFilesystem
When calling the "navigate" method from PermanentFilesystem, save the
result in a cache and try to fetch from cache when looking up the same
folder later.

This cache is not meant to actually reduce the number of network calls,
but rather to provide a somewhat accurate temporary cached value while
the browser is waiting for the API to respond back with the true value.

As a result, the code is designed in such a way that any objects
returned from the `getFolder` method may automatically after the
function is called.

PER-9463: Add caching to folder navigate calls