PER-9494-convert-mixpanel-calls
PER-9494-convert-mixpanel-calls

Created a new Analytics Service which acts as a Subject, injecting the Mixpanel Service into it. This way we can notify all the observers only by importing the analytics service into place. This is a squash and new implementation, because previously I have created a new MixpanelRepo, but this approach made more sense for the future.