Remove unused LoadingArchiveComponent
The LoadingArchiveComponent was created to be an interstitial component
displayed when switching archives in commit 76883bd. However, as time
has gone on, the `/app/switching` path became unused. Since this
specific LoadingArchiveComponent was only used in this one route, the
component itself became unused.

This component has problematic tests. They are (relatively) slow since
they wait 2.5 seconds just for one test, and also sometimes fail due to
random asyncronous errors. While we could optimize the component to fix
these errors, since the component is unused we should just delete it and
the problematic tests.

Resolves PER-9520