Update filePickerClick method
Since this function is called as part of an event, we should make sure
it returns with a proper value so that the rest of the default behavior
is triggered. I'm not sure if it makes a difference, but I also made
this function syncronous since we don't need to await on the results of
the analytics event for anything.

PER-9578: Make upload button work on older browser versions