Use fetch-depth of 0 on upload-sourcemaps action
The Sentry CLI tries to associate a specific group of commits with a
"release" to keep track of when errors happen. Unfortunately, it can't
find all the commits associated with the current version anymore due to
the default depth the checkout action uses when cloning the repository.
Set the fetch-depth to 0 to fetch the whole project history instead of
just a shallow copy.