Create DialogCdkService and Module
The current DialogService used in the web-app uses deprecated Angular
functionality and a few other architectural issues. The new archive type
selection component uses a modal to select a type, so this is a good
opportunity to finally start migrating over to using the Angular CDK
Dialog instead of our custom DialogService. Create a new service that
wraps around Angular CDK's DialogService. The reason to do this right
now is so that there is little confusion over which `DialogService`
components are using, though in the future it would probably be best to
remove this wrapper service and import the Angular CDK DialogService
once we have completely deleted the old one.

See https://github.com/PermanentOrg/web-app/issues/269 for more
information.

PER-9568: Create new component to select archive type