Create unstyled glam-archive-type-select component
Create the pr-glam-archive-type-select component, which will be used to
select the archive type in a future redesigned onboarding flow. This
also includes the archive-type-select-dialog component, which is the
component shown in a modal that is actually what does the selecting.

This commit includes the basic component functionality made through
test-driven development with no styles or accessibility features. These
will come in future commits.

PER-9568: Create new component to select archive type