Add minimize icon and hide button to checklist
Import the minimize icon SVG used in the design and put it in its own
component. Use this icon component, and then add the "Don't show me this
again" button. For now it will do nothing since it needs to send an API
request to fully hide the checklist forever.

PER-9551: Build checklist UI