Update .vscode/settings.json
This project was configured to have various VS Code configuration
included in source control. However, these configuration files haven't
been updated in many years, and so current versions of VS Code try to
update the format automatically. Upgrade the `source.fixAll` parameter
to use the value `"explicit"` instead of `true` to match the current VS
Code config format. Also convert `tslint` to `eslint` since we are using
that now and add support to run prettier on save as well since we are
also using that.