Delete VS Code launch configuration
This repo has included VS Code launch configuration for several years
now. This configuration is used to configure debugging details in the
IDE. Unfortunately, this information is outdated and it seems like there
are not many resources online on how to set up this config for modern
versions of Angular. I think it would be nice to have debugging support
inside the IDE if needed, but I could not get things like remote
debugging working in Karma at all. If we want to add this back in we can
create a dedicated ticket for it, but for now let's at least delete
this obsolete launch config.