Move AccountService calls to ChecklistApi interface
Instead of handing the UserChecklistComponent the whole AccountService,
let's be consistent and only hand it what it needs to know: Does the
current account hide the checklist? We can then put in other methods for
similar calls that will use the AccountService in the background.

PER-9553: Hide checklist if marked hidden in backend