Save account when dismissing checklist
Hide the checklist and attempt to save the account property when
dismissing the new member checklist. If the API call fails, it will fail
silently and the user will just have to try dismissing the checklist
next time they refresh the page.

PER-9553: Hide checklist if marked hidden in backend