Add New Member Checklist to UI
Import the checklist, add it to the main private workspace UI, and
adjust styling so it appears in the corner properly. Despite the fact
that it's now included in the UI by default, the `hideChecklist` account
property is still true by default so this will not show up on production
without additional adjustments.

PER-9571: Integrate checklist into UI