Add method to ChecklistApi for getting Refresh events
The new member checklist should be refreshed when completing events in
the UI. To do that, it needs to be able to subscribe to an external
service. Add a method to the ChecklistApi interface that abstracts these
events so we can separate the logic that handles these events from what
the component is actually subscribed to for testing purposes.

PER-9642: Update checklist when tasks completed