Map analytics events to checklist refresh events
We want to refresh the checklist every time a user completes an action
that potentially checks off a task. We could do this by manually
dispatching refresh events from the various things that check off the
tasks, but in a way that has already been done. Most of the checklist
actions are triggered by analytics events, so we can listen in to the
calls to the analytics service and map them to a Subject which the
checklist component is subscribed to.

The Analytics Service uses an Observer pattern to abstract analytics
calls. We can leverage this by creating a new analytics observer that
just listens for updates and maps them to the Subject that the component
subscribes to

PER-9642: Update checklist when tasks completed