Write Event / Analytics Service Characterization Tests
Write tests that document the current behavior of the Event / Analytics
Services. Right now, we pass in the entire object that is transmitted to
the `event` endpoint in all calls to the EventService. We want to
refactor this so that events are more abstract, and all the
Mixpanel-specific information lives in the AnalyticsService only.

Changing this is a bit risky, so let's create some tests that
characterize the current behavior of the system which can help us make
sure we don't accidentally break anything as we refactor. The tests
right now are very simple, since we're only asserting that the object
passed in is the same one that is sent to the endpoint in almost all
cases. But as we change the call to `event.notifyObservers`, these tests
will become critical to ensure that our abstract event calls still send
out identical bodies to the actual `/event` endpoint.

PER-9642: Update checklist when tasks completed