Refactor EventService to only take PermanentEvent
Make event calls more simple and abstract by only expecting
PermanentEvent objects instead of the large Mixpanel-focused analytics
objects. Rewrite the AnalyticsService to take these simple event objects
and adapt them to the Mixpanel-specific format that we currently use,
and ignore events it does not have configuration for. Then, update all
calls to the EventService's `notifyObservers` method.

Some of this refactoring also meant moving around some of the
responsibility of certain functionality. Instead of having to pass the
AccountId into every single account event, the AnalyticsService instead
has an injected AccountService it can use to get the current account
value. The AnalyticsService also has taken ownership of the code that
checks Mobile Width and Public/Private Workspaces. A future commit will
delete the `DeviceService.getViewMessageForEventTracking` method.

Deleting some of the old types and other cleanup tasks will be done in a
future commit as well.

The characterization tests now replace the old AnalyticsService tests
since the AnalyticsService was pretty much rewritten and are a good way
to cover all of the events anyway.

PER-9642: Update checklist when tasks completed