Remove old `EventData` types and usages
Delete the old EventData type now that it has been replaced by
`PermanentEvent`. Also remove all imports of it. In the AnalyticsService
integration tests, these references are replaced with the MixpanelData
type that is still used internally. Other reference to the EventData
type can safely be deleted.

Many tests referencing the `EventData` or `PermanentEvent` types were
mocking out the AnalyticsService. Previously this was necessary since
the original `MixpanelService` would send out requests or interact with
the Mixpanel library directly; something we really didn't want to do in
isolated component or service unit tests.

Since the way we dispatch analytics events has changed a while back with
the usage of the Observer pattern, we can safely use the real
`EventsService` in tests without having to mock it out at all. This is
because by default, the EventsService does nothing until it actually has
observers listening to its events. Delete the mocked EventService
implementations throughout the codebase as well since they're not needed
anymore.

PER-9642: Update checklist when tasks completed