Rename EventService.notifyObservers to dispatch
The name "NotifyObservers" describes what it's doing internally
(notifying all of the EventService's observers) but it's quite generic
and a bit verbose. In an editor with some kind of "intellisense"
functionality, it might be difficult to figure out which method of
EventService actually sends out events (the most important method of the
whole class!).

Since this class and method deal with broadcasting events to interested
observers, I think the best name is "dispatch". It is simple, short, and
seems to be the right verb to use with event objects.

PER-9642: Update checklist when tasks completed