Rename references to AnalyticsService
The typical component property name assigned to the old AnalyticsService
was `analytics`. We've since renamed this service to the EventService,
so it only makes sense to rename any names that reference analytics and
replace it with `event`. This would make the typical method call to
dispatch an event `this.event.dispatch(...)` which is much more readable
than `this.analytics.notifyObservers(...)`. While in these various
component/service files, I have also done some additional cleanup work:
removing unused imports and unused component properties (usually a
DeviceService or AccountService import that was used in previous
analytics calls that are no longer used now).

Also rename the ChecklistAnalyticsObserverService to
ChecklistEventObserverService. (Right, that was what sparked all these
refactorings!)

PER-9642: Update checklist when tasks completed