Provide DataService and Event Services at app level
The AnalyticsService now depends on the DataService. Since it is
initialized in the AppModule, provide the DataService in the top
AppModule now and remove it from the CoreModule's providers so that
there is only one instance of the DataService being shared between all
components and services.

PER-9642: Update checklist when tasks completed