Add delay to ChecklistEventObserverService
Add a configurable delay to the ChecklistEventObserverService. Since
event dispatches are not `await`ed on, calls to the event endpoint and
refreshing the checklist happen at pretty much the same time. Add a
delay to the checklist event observer so that there is time for events
to hit the endpoint before we refresh the checklist. Make it
configurable so we are not waiting that much time in the unit tests.

PER-9642: Update checklist when tasks completed