Remove CSS url imports
CSS imports create blocking HTTP requests. Styles in the same sheet are
not applied until after the external resource loads, and if the resource
does not load it seems like browsers do not apply the rules of the whole
stylesheet. This can cause issues if a user has a spotty connection, is
blocking one of the external CDNs we are using, or is located somewhere
where they are blocked by the CDN.

An easy alternative to this is just to include the imported CSS as a
normal stylesheet in the HTML. This is fetched asynchronously, and if it
fails it does not prevent other stylesheets from loading or applying.

PER-9726: Styles are completely broken if some external resources don't
load