Remove "skip step" link on create archive screen
The logic to hide this link on the create archive screen was
accidentally removed when integrating some hidden GLAM redesign work in
this component. Restore it so that there isn't a weird broken link on
onboarding. Also add logic to hide on the "name-archive" step which
previously didn't exist, but is now (sometimes) used in the old flow
because of some of the GLAM code.

PER-9742: Hidden GLAM onboarding bugs