Move create archive buttons to original position
These buttons were moved around the create archive component as a result
of some of the hidden GLAM work. As a result, they are covered by an
overlay that hides the Archive Name form that is present when the user
hasn't selected an archive type yet. We want the "Make my archive for
me" and (in the case of the archive member flow) "Go back" buttons to be
visibly clickable, so move all these back to their original positions.

PER-9742: Hidden GLAM onboarding bugs