Extract out onboarding header component
We want to create a new onboarding header for the GLAM redesign. To make
sure that the old header is insulated from any GLAM changes, extract it
out as its own component. Then we can just switch between the two
components depending on whether Glam redesignes are enabled or not.

PER-9753: Glam Onboarding Header