Create Onboarding Demo in Storybook
The onboarding flow is a bit hard to test styles of since the
development team needs to constantly create new accounts on a running
Permanent environment. Originally when the Onboarding flow was under
construction, we simulated network calls and controlled this with a
debugging component embedded into onboarding. Now we have access to
something even more powerful: Storybook.

Create a set of stories in Storybook that allow us to test a mocked
version of onboarding without having to connect to the actual API.
Included are stories for the default "classic" onboarding flow, the
"classic" flow with pending archive invitations, and the new GLAM flow.

This should hopefully simplify making style changes to the onboarding
flow.

PER-9753: Glam Onboarding Header