Call v2 version of `account/update` endpoint
Call the v2 version of the `account/update` endpoint instead of the v1
version. This changes the signature of the account.update method of the
ApiService, since it no longer returns a ResponseVO. Update calls to
`account.update` and also do some slight refactoring.

Both the BillingSettingsComponent and the AccountSettingsComponent have
identical logic when it comes to saving account properties. Extract this
logic into a helper function so we can deduplicate this logic but still
operate on our component properties and services directly.

PER-9747: Use a v2 request when updating hideChecklist