Create GetThumbnail function
This function gets the thumbnail of a record or folder that matches a
specific minimum size. The intention is to generalize thumbnail logic so
that when we start generating 256x256 thumbnails, all code using old
thumbnails can switch over to them without any issues. This is a
function so that it can be used by RecordVOs and FolderVOs.

PER-9671: Update web-app to handle new thumbnail sizes