Refactor ThumbnailComponent tests and add coverage
The ThumbnailComponent needs to be modified to be able to handle new
thumbnail sizes and preferably use the new `GetThumbnail` function we
want to use. Before we do that, let's cover as much thumbnail
functionality as we can with tests so we know we're not modifying any
important behavior.

Use the shallow render library in the thumbnail component tests, provide
methods to give us access to private properties in tests, and add new
tests that cover folders and image loading.

Previously the test thumbnail URLs were pointing to a free placeholder
image generator service. This placeholder service is no longer up at the
old URL, so point to a different one instead. Also provide a mock for
the `Image` class so we can reliably test loading images without having
to actually worry about the domain going down, or even being connected
to the internet in the first place. This does mean we have to provide
the default Image class in the main SharedModule.

PER-9671: Update web-app to handle new thumbnail size(s)