Refactor ThumbnailComponent to use GetThumbnail
Now that we have tests covering the ThumbnailComponent, refactor it to
use the `GetThumbnail` function for fetching thumbnails. This lets us
clean up a lot of the component logic and delete a lot of unnecessary
code. Clean up some redundant code and also remove usage of the
`Renderer2` object and replace them with directives in the HTML
template.

PER-9671: Update web-app to handle new thumbnail size(s)