Extract ZoomingImageViewer from ThumbnailComponent
Originally, the web-app displayed image files by showing a record's
biggest thumbnail (2000px). As a result, the `pr-thumbnail` component
was used in the file viewer. When we originally added support for
zooming in on images with OpenSeaDragon, this zooming functionality was
added to that `pr-thumbnail` component. This is not ideal, because
components shouldn't have multiple responsibilities like this. We use
the `pr-thumbnail` component in other places in the app, and it would be
best if there were not unused OpenSeaDragon functionalities lurking
within all of those other thumbnails. The coupling between these two
functions also just made it hard to parse and test the `pr-thumbnail`
component when refactoring it to use the `GetThumbnail` function.

Create a new ZoomingImageViewerComponent with tests, remove the
OpenSeaDragon functionality from the ThumbnailComponent, and use this
ZoomingImageViewerComponent in the file viewer instead of the
ThumbnailComponent when necessary.

For some reason, OpenSeaDragon runs very slowly in unit tests, even
though in theory we should be able to trigger all events manually in
test code. To address this issue, OpenSeaDragon is mocked in unit tests
and a real copy of it is provided in the SharedModule.

PER-9671: Update web-app to handle new thumbnail size(s)