Extract redeem-gift component from storage dialog
The tabbed modals we use for various settings throughout the app tend to
bundle all of the tab logic together. Ideally, all of these tabs should
be isolated components. This makes them easier to test and maintain.
Extract the redeem-gift component from the storage dialog, and make sure
that each component owns only what it needs to.

Also add a storybook story for this new component so we can test this
easily.

PER-9750: Add gift code redemption success/fail message to modal