Log out user with expired JWT on V1 requests
Some V1 endpoints require a JWT now, and will return with an error code
if the token is expired. Log the user out if such an error code is
returned similar to how we do with the HttpV2 service.

PER-9702: Endpoints can fail with an expired JWT with no log out