Fix "View On Web" redirect on Profile Settings
Fix an issue that occurs when navigating away from a routed dialog
component. Specifically, the `closedByNavigate` variable is not set
until after the dialog is closed meaning it is not set when it is
needed. Fix an additional issue that can cause a loop when clicking the
"View On Web" button and then hitting the back button by removing
popstate events from the route history service.

At some point we should reconsider how we're currently dealing with
dialogs, route history, and the routed dialog wrapper. The current
system does not seem to be designed with the idea that links may be
contained within dialogs.

PER-9789: "View on Web" redirects to archive manager