Do not update account storage if it's already updated
When redeeming a promo code, the web-app updates the account client side
to reflect the new storage the user has. However it also fetches an
updated version of the account from the API, which means it may already
get that storage update from the server. Update the logic to only run if
the changes are not reflected in the server response.

PER-9798: Using promo code results in 2x space than expected