Load archive information properly on login
The default (non-MFA) login process does not load up to date archive
information as part of its flow. As a result, all logins currently
redirect the user to onboarding since the login code doesn't see any
archives that belong to the user. Thankfully this doesn't actually
create any major issues, since the onboarding page immediately runs a
check for archives and redirects the user back to the private workspace.

This does create a few subtle issues though:
  - The default archive is not being set in a way that the TagsService
    can notice it
  - The "Logged in as $user" message is not being displayed

The former issue in particular means that tags are not loaded properly
in a standard login. Add a call to "refreshArchives" in the login
process so we can properly run the "non-onboarding non-MFA" login
routine, and then add a call to `switchToDefaultArchive` in that routine
since we know the user has at least one non-pending archive that must be
a default. The latter function calll also triggers an event that causes
the TagsService to load all archive tags correctly.

PER-9810: Archive tags are not loaded on login