Only make one `getTagsByArchive` API call
The TagsService would unintentionally call the `refreshTags` function
twice whenever the user's archive is changed or loaded for the first
time. Remove the unnecessary call so that we're only hitting the
endpoint once (as of this commit it's still a PHP V1 endpoint, so this
is costing us a tiny bit of performance from other V1 requests).

PER-9810: Archive tags are not loaded on login