Fix SidebarComponent tests
A Pipe was being mocked by shallow-render when we didn't want it to be.
Add a `dontMock` section with the `asRecord` pipe provided. Also add a
null check to `selectedItem.FileVOs`, which seemed to be triggering
issues in tests. Clean up some of the code formatting and unnecessary
lines in the SidebarComponent tests while we're there.