Add FeatureFlagModule and FeatureFlagService
Add a service designed to query feature flags that are fetched from the
API. This behavior is implemented in two classes: one for storing and
querying the currently set flags and one for initially fetching them
from the API.

It is designed like this so that we can test these functionalities
on their own and so that the FeatureFlagService can be used in unit tests
without any mocking. Any components that query FeatureFlags can just use
the service directly and use the `set` method to manipulate feature
flags directly in the test.

PER-9735: Create service to fetch client config