Fix InlineValueEditComponent test failure
This test would sometimes randomly fail, always as a result of the
`onTextInputBlur` spy not registering a call sometimes. An easy fix to
this test is just to remove the check for the call to `onTextInputBlur`.
We do not need this test to be *that* specific, since we are already
verifying that the `doneEditing` output is not being emitted and there
is no additional logic that happens in the `onTextInputBlur` method. In
general, we should avoid having specific method spies in our tests, as
it couples the tests too much to the internal design of the component it
is testing. If we refactored this component or even renamed the
`onTextInputBlur` method, the test would fail despite no actual behavior
changes. That would be bad, so let's just get rid of this spy and
assertion completely.

PER-9898: InlineValueEditComponent test failure