Fix ArchiveCreationWithShareComponent tests
These tests had some issues with mocking. Provide the same mock to each
test so we don't have to mock out HTTP response information for each
test, and also use some more abstract logic in our assertions instead of
using a spy on our mock.

To make some of these tests work, the InviteVoData interface needed to
be modified to accept `AccountVO`, `RecordVO`, and `FolderVO`
properties. These were already being passed into values that were
supposed to be of type `InviteVoData` so this does not actually change
any logic and it should have been like this to begin with.

PER-9900: ArchiveCreationWithShareToken test failures