Fix "an error was thrown in afterAll" test failure
This elusive test failure could sometimes be seen in random tests
throughout the web-app, always with the unhelpful message that
`accountId` is undefined but providing no actual context of which test
is actually failing.

The current lead is that this test failure is caused by the
`AccountService.setAccount` method. This method also ends up configuring
the current Sentry scope so that the current user ID can be logged by
Sentry. There is no null check for this action, so tests where we
nullify the current account are subject to possibly fail.

Add a null check around this Sentry scope check. Ideally this Sentry
code should probably live outside of the AccountService so that we can
skip running it in unit tests to begin with.

PER-9592: An error was thrown in afterAll