Centralize download list logic into RecordVO
Add a method onto RecordVO that gets a list of info about download
options from its list of FileVOs. This function is duplicated across the
codebase for each of the download button components so adding this as a
method onto RecordVO centralizes this logic in one place. Then use this
method in both download buttons.

PER-9902: Include Archivematica access copies in the download options