Add `prioritizeIf` helper function
When sorting through FileVOs a hard-to-read pattern emerged multiple
times that involved subtracting booleans that were converted to numbers.
Extract this logic into its own helper function so we can continue to
use this helpful pattern but it also is more readable and more reusable.

PER-9902: Include Archivematica access copies in the download options