Create OnboardingService
The onboarding flow(s) used in the app tend to involve a lot of small
components passing data to each other. This communication can be simple
and lightweight, but as components and flows grow in complexity this
becomes harder to manage. Fortunately we can use Angular services as a
way to refactor this code. Create the OnboardingService that stores
created/accepted archives so we can show them back to the user in the
final screen without having to differentiate between created and
accepted archives.

While it's unclear if just this refactoring requires a whole service,
there have been many occurences where consolidating onboarding code into
a central service would have been useful in the past few months. So
while the implementation of this service leaves a lot to be desired,
this is at least a good first step towards moving more of the onboarding
flow's behavior into this service instead of components. This in turn
would allow us to redesign the onboarding flow in the future without
worrying about damaging fundamental onboarding behavior.

PER-9891: Final glam onboarding screen should list all accepted archives