Get final archives from OnboardingService
Get finalized archives from the OnboardingService to display in the
`pr-finalize-archive-creation-screen` component and update the component
to iterate over a list of archives instead of just displaying the name.
This also means updating the component so access role is not hardcoded
to list "OWNER".

PER-9891: Final glam onboarding screen should list all accepted archives