Remember accepted archives in glam onboarding
If a user hits the "approve" button, use the new OnboardingService to
remember this if the user hits back during the process. Also, adjust how
archives are added to the service so that a user can accept archives AND
create a new archive, and all archives are present at the end of the
list.

PER-9891: Final glam onboarding screen should list all accepted archives