Add scrollbar for long list of FamilySearch tree members
Previously this probably scrolled but with the new dialog changes
scrolling needed to be added to the dialog. To make things flow better,
add a scroll bar in the middle of the dialog so that the prompt ("select
family members") and the buttons are always visible while a user is
choosing something.

PER-9917: Fix FamilySearch Modal