Check environment variables in FeatureFlagService
This is mainly for use in development environments, where setting
feature flags in local Stela can be a bit annoying. A secret loaded from
the `.env` file can be used in place of fetching a feature flag from the
back-end. It will not overwrite a value from the back-end so this is
only a fallback method to set values not defined in Stela already.

Also add the ability to store secrets with kebab-case formats.

To actually get this to work, a dev will have to temporarily store the
feature flag in the "optional secrets" file. I do not recommend actually
comitting that change though.

PER-9570: Integrate Glam Onboarding Feature Flag